Recipe結合テスト用ファイルの作成
・デフォルトのExampleTestへ直書きしていたRecipe関連の結合テストをFeature/RecipeTestに切り出しました。
