+++ code/tests/Feature/ExampleTest.php
@@ -9,7 +9,7 @@ use App\Recipe;
 class ExampleTest extends TestCase
 {
     /** 
-     * 事前マイグレートとRecipeのレコード作成
+    * 事前マイグレートとRecipeのレコード作成
     */
     protected function setUp(): void
     {
@@ -35,40 +35,4 @@ class ExampleTest extends TestCase
         $response->assertStatus(200);
     }
 
-    /**
-     * 指定したidのRecipeレコードがある時、レシピ詳細ページが正常に表示されるかどうかのテスト
-     *
-     * @return void
-     */
-    public function testShowRecipeIsAvailableWithRecord()
-    {
-        $recipe = Recipe::latest()->first();
-        $response = $this->get('/recipes/'.$recipe->id);
-
-        $response->assertStatus(200);
-    }
-
-    /**
-     * 指定したidのRecipeレコードがない時、レシピ詳細ページが正常に表示されないことのテスト
-     *
-     * @return void
-     */
-    public function testShowRecipeIsNotAvailableWithoutRecord()
-    {
-        $recipe = Recipe::latest()->first();
-        $response = $this->get('/recipes/999');
-
-        $response->assertStatus(404);
-    }
-
-
-    /**
-     * Recipeのサンプルオブジェクトの生成
-     */
-    private static function newSampleRecipe(){
-        $recipe = new Recipe();
-        $recipe->title = "サンプルレシピ";
-        $recipe->description = "サンプル説明文";
-        return $recipe;
-    }
 }
